Move chat related files in backend
