+++ backend/src/routes/public/chat.ts
@@ -4,7 +4,7 @@ import {
   ChatCompletionResponse,
 } from "gpinterface-shared/type/chat";
 import { Static } from "@sinclair/typebox";
-import { createChatCompletion } from "../../chat/controllers/chat";
+import { createChatCompletion } from "../../controllers/chat";
 import { GpiHashIdParam } from "gpinterface-shared/type";
 
 export default async function (fastify: FastifyInstance) {
